api/requirements.txt
Django==3.2.4
pytest==6.2.3
pytest-cov ==2.11.1
pytest_django==4.1.0
django-autoslug==1.9.8
django-model-utils==4.1.0
pillow==8.0.1
factory-boy==3.1.0
django_environ==0.4.5
whitenoise==5.2.0
psycopg2==2.9.5
gunicorn==20.0.4
django-storages==1.13.2
boto3==1.26.67